+++ app/src/main/AndroidManifest.xml
@@ -2,11 +2,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="seniorproject.badger">
 
-    <!-- To auto-complete the email text field in the login form with the user's emails -->
-    <uses-permission android:name="android.permission.GET_ACCOUNTS" />
-    <uses-permission android:name="android.permission.READ_PROFILE" />
-    <uses-permission android:name="android.permission.READ_CONTACTS" />
-
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -25,14 +20,6 @@
         <activity
             android:name=".Profile"
             android:label="@string/title_activity_profile" />
-        <activity
-            android:name=".LoginCredentialsActivity"
-            android:label="@string/title_activity_login_credentials"
-            android:parentActivityName=".Login">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="seniorproject.badger.Login" />
-        </activity>
         <activity
             android:name=".Badge"
             android:label="@string/title_activity_badge" />
@@ -58,6 +45,24 @@
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="seniorproject.badger.Profile" />
+        </activity>
+        <activity
+            android:name=".CreateAccount"
+            android:label="@string/title_activity_create_account"
+            android:parentActivityName=".LoginCredentials"
+            android:theme="@style/AppTheme">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value="seniorproject.badger.LoginCredentials" />
+        </activity>
+        <activity
+            android:name=".LoginCredentials"
+            android:label="@string/title_activity_login_credentials"
+            android:parentActivityName=".Login"
+            android:theme="@style/AppTheme">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value="seniorproject.badger.Login" />
         </activity>
     </application>
